Use FieldList by default
The adder method is now accessed by scheme.fields.add instead of scheme.add_field as long as you don't overwrite the fields attribute with a regular list. Later we could make the fields attribute a property and disallow setting it to anything but an FieldList, but that feels like overkill for now.